+++ tests/observables/test_rx_create.py
@@ -11,7 +11,8 @@ def test_rx_create_profile_test():
         rx_create(subscribe=None)  # type: ignore
 
 
-def test_rx_create_with_default(kernel):
+@pytest.mark.curio
+async def test_rx_create_with_default():
     async def _subscribe(an_observer: Observer):
         await an_observer.on_next(item=1)
         await an_observer.on_completed()
@@ -25,15 +26,16 @@ def test_rx_create_with_default(kernel):
     assert seeker.on_completed_count == 0
     assert seeker.on_error_count == 0
 
-    unsub = kernel.run(obs.subscribe(seeker))
-    kernel.run(unsub())
+    unsub = await obs.subscribe(seeker)
+    await unsub()
 
     assert seeker.on_next_count == 1
     assert seeker.on_completed_count == 1
     assert seeker.on_error_count == 0
 
 
-def test_rx_create_with_no_ensure_contract(kernel):
+@pytest.mark.curio
+async def test_rx_create_with_no_ensure_contract():
     async def _subscribe(an_observer: Observer):
         await an_observer.on_next(item=1)
         await an_observer.on_completed()
@@ -47,15 +49,16 @@ def test_rx_create_with_no_ensure_contract(kernel):
     assert seeker.on_completed_count == 0
     assert seeker.on_error_count == 0
 
-    unsub = kernel.run(obs.subscribe(seeker))
-    kernel.run(unsub())
+    unsub = await obs.subscribe(seeker)
+    await unsub()
 
     assert seeker.on_next_count == 1
     assert seeker.on_completed_count == 1
     assert seeker.on_error_count == 0
 
 
-def test_rx_create_max_observer(kernel):
+@pytest.mark.curio
+async def test_rx_create_max_observer():
     async def _subscribe(an_observer: Observer):
         await an_observer.on_next(item=1)
         await an_observer.on_completed()
@@ -69,16 +72,16 @@ def test_rx_create_max_observer(kernel):
     assert seeker.on_completed_count == 0
     assert seeker.on_error_count == 0
 
-    unsub = kernel.run(obs.subscribe(seeker))
-    kernel.run(unsub())
+    unsub = await obs.subscribe(seeker)
+    await unsub()
 
     assert seeker.on_next_count == 1
     assert seeker.on_completed_count == 1
     assert seeker.on_error_count == 0
 
     # ok because we had called unsub
-    unsub = kernel.run(obs.subscribe(seeker))
+    unsub = await obs.subscribe(seeker)
 
     # raise because max_observer = 1
     with pytest.raises(RuntimeError):
-        unsub = kernel.run(obs.subscribe(seeker))
+        unsub = await obs.subscribe(seeker)