+++ tests/observables/test_rx_empty.py
@@ -6,7 +6,8 @@ from async_rx.protocol import default_subscription
 from ..model import ObserverCounter
 
 
-def test_rx_empty(kernel):
+@pytest.mark.curio
+async def test_rx_empty():
 
     obs: Observable = rx_empty()
 
@@ -16,8 +17,8 @@ def test_rx_empty(kernel):
     assert seeker.on_completed_count == 0
     assert seeker.on_error_count == 0
 
-    unsub = kernel.run(obs.subscribe(seeker))
-    kernel.run(unsub())
+    unsub = await obs.subscribe(seeker)
+    await unsub()
 
     assert seeker.on_next_count == 0
     assert seeker.on_completed_count == 1