+++ tests/observables/test_rx_forward.py
@@ -19,7 +19,8 @@ async def _subscribe_with_error(an_observer: Observer):
     return default_subscription
 
 
-def test_rx_forward_with_default(kernel):
+@pytest.mark.curio
+async def test_rx_forward_with_default():
     obs: Observable = rx_create(subscribe=_subscribe)
 
     forwarded_obs = rx_forward(observable=obs)
@@ -29,15 +30,16 @@ def test_rx_forward_with_default(kernel):
     assert seeker.on_completed_count == 0
     assert seeker.on_error_count == 0
 
-    unsub = kernel.run(forwarded_obs.subscribe(seeker))
-    kernel.run(unsub())
+    unsub = await forwarded_obs.subscribe(seeker)
+    await unsub()
 
     assert seeker.on_next_count == 1
     assert seeker.on_completed_count == 1
     assert seeker.on_error_count == 0
 
 
-def test_rx_forward_without_complete(kernel):
+@pytest.mark.curio
+async def test_rx_forward_without_complete():
     obs: Observable = rx_create(subscribe=_subscribe)
 
     forwarded_obs = rx_forward(observable=obs, except_complet=True)
@@ -47,15 +49,16 @@ def test_rx_forward_without_complete(kernel):
     assert seeker.on_completed_count == 0
     assert seeker.on_error_count == 0
 
-    unsub = kernel.run(forwarded_obs.subscribe(seeker))
-    kernel.run(unsub())
+    unsub = await forwarded_obs.subscribe(seeker)
+    await unsub()
 
     assert seeker.on_next_count == 1
     assert seeker.on_completed_count == 0  # not forwarded
     assert seeker.on_error_count == 0
 
 
-def test_rx_forward_without_error(kernel):
+@pytest.mark.curio
+async def test_rx_forward_without_error():
 
     seeker = ObserverCounter()
     assert seeker.on_next_count == 0
@@ -64,22 +67,23 @@ def test_rx_forward_without_error(kernel):
 
     # no error because is silent
     with pytest.raises(Exception):
-        kernel.run(rx_create(subscribe=_subscribe_with_error).subscribe(seeker))
+        await rx_create(subscribe=_subscribe_with_error).subscribe(seeker)
     assert seeker.on_next_count == 1
     assert seeker.on_completed_count == 0  # because ensure contract enabled
     assert seeker.on_error_count == 1
 
-    kernel.run(rx_forward(observable=rx_create(subscribe=_subscribe_with_error), except_error=True).subscribe(seeker))
+    await rx_forward(observable=rx_create(subscribe=_subscribe_with_error), except_error=True).subscribe(seeker)
     assert seeker.on_next_count == 2
     assert seeker.on_completed_count == 0  # because ensure contract enabled
     assert seeker.on_error_count == 1
 
 
-def test_rx_forward_without_completed(kernel):
+@pytest.mark.curio
+async def test_rx_forward_without_completed():
 
     seeker = ObserverCounter()
 
-    kernel.run(rx_forward(observable=rx_create(subscribe=_subscribe), except_complet=True).subscribe(seeker))
+    await rx_forward(observable=rx_create(subscribe=_subscribe), except_complet=True).subscribe(seeker)
     assert seeker.on_next_count == 1
     assert seeker.on_completed_count == 0
     assert seeker.on_error_count == 0