+++ tests/observables/test_rx_from.py
@@ -6,20 +6,22 @@ from async_rx.protocol import default_subscription
 from ..model import ObserverCounterCollector
 
 
-def test_rx_from_dict(kernel):
+@pytest.mark.curio
+async def test_rx_from_dict():
     result = rx_from({"A": True})
     assert hasattr(result, "subscribe")
     result["B"] = False
     assert result == {"A": True, "B": False}
 
 
-def test_rx_from_iterable(kernel):
+@pytest.mark.curio
+async def test_rx_from_iterable():
 
     obs: Observable = rx_from(observable_input=[1, 2, 3])
 
     seeker = ObserverCounterCollector()
 
-    kernel.run(obs.subscribe(seeker))
+    await obs.subscribe(seeker)
 
     assert seeker.on_next_count == 3
     assert seeker.on_completed_count == 1
@@ -27,7 +29,8 @@ def test_rx_from_iterable(kernel):
     assert seeker.items == [1, 2, 3]
 
 
-def test_rx_from_observable(kernel):
+@pytest.mark.curio
+async def test_rx_from_observable():
     async def _subscribe(an_observer: Observer):
         await an_observer.on_next(item=1)
         await an_observer.on_completed()
@@ -36,7 +39,7 @@ def test_rx_from_observable(kernel):
     obs: Observable = rx_from(observable_input=rx_create(subscribe=_subscribe))
     seeker = ObserverCounterCollector()
 
-    kernel.run(obs.subscribe(seeker))
+    await obs.subscribe(seeker)
 
     assert seeker.on_next_count == 1
     assert seeker.on_completed_count == 1
@@ -44,7 +47,8 @@ def test_rx_from_observable(kernel):
     assert seeker.items == [1]
 
 
-def test_rx_from_slug(kernel):
+@pytest.mark.curio
+async def test_rx_from_slug():
     class Temp:
         async def subscribe(self, an_observer: Observer):
             await an_observer.on_next(item=1)
@@ -54,7 +58,7 @@ def test_rx_from_slug(kernel):
     obs: Observable = rx_from(observable_input=Temp())
     seeker = ObserverCounterCollector()
 
-    kernel.run(obs.subscribe(seeker))
+    await obs.subscribe(seeker)
 
     assert seeker.on_next_count == 1
     assert seeker.on_completed_count == 1
@@ -62,12 +66,13 @@ def test_rx_from_slug(kernel):
     assert seeker.items == [1]
 
 
-def test_rx_from_singleton(kernel):
+@pytest.mark.curio
+async def test_rx_from_singleton():
 
     obs: Observable = rx_from(observable_input=42)
     seeker = ObserverCounterCollector()
 
-    kernel.run(obs.subscribe(seeker))
+    await obs.subscribe(seeker)
 
     assert seeker.on_next_count == 1
     assert seeker.on_completed_count == 1
@@ -75,7 +80,8 @@ def test_rx_from_singleton(kernel):
     assert seeker.items == [42]
 
 
-def test_rx_from_awaitable_iterable(kernel):
+@pytest.mark.curio
+async def test_rx_from_awaitable_iterable():
     async def generate():
         for i in range(4):
             yield i
@@ -83,7 +89,7 @@ def test_rx_from_awaitable_iterable(kernel):
     obs: Observable = rx_from(observable_input=generate())
     seeker = ObserverCounterCollector()
 
-    kernel.run(obs.subscribe(seeker))
+    await obs.subscribe(seeker)
 
     assert seeker.on_next_count == 4
     assert seeker.on_completed_count == 1