+++ tests/observables/test_rx_merge_map.py
@@ -8,7 +8,8 @@ from ..model import ObserverCounterCollector
 from .model import countdown
 
 
-def test_rx_merge_map(kernel):
+@pytest.mark.curio
+async def test_rx_merge_map():
 
     seeker = ObserverCounterCollector()
 
@@ -18,16 +19,17 @@ def test_rx_merge_map(kernel):
     async def _build():
         return rx_merge_map(rx_create(subscribe=await countdown(2, 0.1)), rx_create(subscribe=await countdown(2, 0.2)), transform=adder)
 
-    obs = kernel.run(_build())
-    sub_a = kernel.run(obs.subscribe(seeker))
-    kernel.run(sub_a())
+    obs = await _build()
+    sub_a = await obs.subscribe(seeker)
+    await sub_a()
     assert seeker.on_completed_count == 1
     assert seeker.on_error_count == 0
     assert seeker.on_next_count == 4
     assert seeker.items == [3, 3, 2, 2]
 
 
-def test_rx_merge_map_sync(kernel):
+@pytest.mark.curio
+async def test_rx_merge_map_sync():
 
     seeker = ObserverCounterCollector()
 
@@ -37,9 +39,9 @@ def test_rx_merge_map_sync(kernel):
     async def _build():
         return rx_merge_map(rx_create(subscribe=await countdown(2, 0.1)), rx_create(subscribe=await countdown(2, 0.2)), transform=adder)
 
-    obs = kernel.run(_build())
-    sub_a = kernel.run(obs.subscribe(seeker))
-    kernel.run(sub_a())
+    obs = await _build()
+    sub_a = await obs.subscribe(seeker)
+    await sub_a()
     assert seeker.on_completed_count == 1
     assert seeker.on_error_count == 0
     assert seeker.on_next_count == 4