+++ tests/subject/test_replay_subject.py
@@ -13,7 +13,8 @@ def test_replay_subject_default():
         rx_subject_replay(-1)
 
 
-def test_replay_subject(kernel):
+@pytest.mark.curio
+async def test_replay_subject():
 
     seeker_a = ObserverCounterCollector()
     seeker_b = ObserverCounterCollector()
@@ -22,23 +23,24 @@ def test_replay_subject(kernel):
     assert a_subject
 
     # first registration
-    sub_a = kernel.run(a_subject.subscribe(seeker_a))
-    sub_subject = kernel.run(rx_range(start=0, stop=10).subscribe(a_subject))
+    sub_a = await a_subject.subscribe(seeker_a)
+    sub_subject = await rx_range(start=0, stop=10).subscribe(a_subject)
     assert seeker_a.on_next_count == 10
     assert seeker_a.on_error_count == 0
     assert seeker_a.on_completed_count == 1
 
     # second registration
-    sub_b = kernel.run(a_subject.subscribe(seeker_b))
+    sub_b = await a_subject.subscribe(seeker_b)
     assert seeker_b.on_next_count == 4  # buffer size
     assert seeker_b.on_error_count == 0
     assert seeker_b.on_completed_count == 1
 
-    kernel.run(sub_a())
-    kernel.run(sub_b())
+    await sub_a()
+    await sub_b()
 
 
-def test_replay_subject_with_error(kernel):
+@pytest.mark.curio
+async def test_replay_subject_with_error():
     async def _subscribe(an_observer: Observer) -> Subscription:
         await an_observer.on_next("A")
         await an_observer.on_error("Args")
@@ -48,20 +50,20 @@ def test_replay_subject_with_error(kernel):
     seeker_b = ObserverCounterCollector()
     a_subject = rx_subject_replay(buffer_size=4)
 
-    sub_a = kernel.run(a_subject.subscribe(seeker_a))
+    sub_a = await a_subject.subscribe(seeker_a)
 
-    sub_subject = kernel.run(rx_create(subscribe=_subscribe).subscribe(a_subject))
+    sub_subject = await rx_create(subscribe=_subscribe).subscribe(a_subject)
 
     assert seeker_a.on_next_count == 1
     assert seeker_a.on_error_count == 1
     assert seeker_a.on_completed_count == 0
     assert seeker_a.items == ["A"]
 
-    sub_b = kernel.run(a_subject.subscribe(seeker_b))
+    sub_b = await a_subject.subscribe(seeker_b)
     assert seeker_b.on_next_count == 1
     assert seeker_b.on_error_count == 1
     assert seeker_b.on_completed_count == 0
     assert seeker_b.items == ["A"]
 
-    kernel.run(sub_a())
-    kernel.run(sub_b())
+    await sub_a()
+    await sub_b()