+++ tests/subject/test_subject_behavior.py
@@ -1,10 +1,13 @@
+import pytest
+
 from async_rx import rx_range, rx_subject_behavior
 from async_rx.protocol import Observer, subject_handler
 
 from ..model import ObserverCounterCollector
 
 
-def test_replay_subject(kernel):
+@pytest.mark.curio
+async def test_replay_subject():
 
     current = None
 
@@ -19,19 +22,19 @@ def test_replay_subject(kernel):
     assert a_subject
 
     # first registration
-    sub_a = kernel.run(a_subject.subscribe(seeker_a))
-    sub_subject = kernel.run(rx_range(start=0, stop=10).subscribe(a_subject))
+    sub_a = await a_subject.subscribe(seeker_a)
+    sub_subject = await rx_range(start=0, stop=10).subscribe(a_subject)
     assert seeker_a.on_next_count == 10
     assert seeker_a.on_error_count == 0
     assert seeker_a.on_completed_count == 1
     assert current == 1
     # second registration
-    sub_b = kernel.run(a_subject.subscribe(seeker_b))
+    sub_b = await a_subject.subscribe(seeker_b)
     assert seeker_b.on_next_count == 1  # buffer size
     assert seeker_b.on_error_count == 0
     assert seeker_b.on_completed_count == 1
     assert current == 2
-    kernel.run(sub_a())
+    await sub_a()
     assert current == 1
-    kernel.run(sub_b())
+    await sub_b()
     assert current == 0